cdk/requirements.txt
aws-cdk-lib==2.133.0
constructs>=10.0.0
